requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pymysql>=1.1.0  # For MySQL/MariaDB
psycopg2-binary>=2.9.0  # For PostgreSQL
schedule>=1.2.0
pydantic>=2.0.0
python-dateutil>=2.8.0
